+++ app/src/main/java/com/example/streams/PlayerActivity.kt
@@ -9,9 +9,9 @@ import com.tpstream.player.*
 
 class PlayerActivity : AppCompatActivity() {
     lateinit var playerFragment: TpStreamPlayerFragment;
-    private val accessToken = "bbf23112-0c14-4519-a848-73c95cb024ac"
-    private val videoId = "E44ulfSWhYx"
-    private val orgCode = "drm"
+    private val accessToken = "c381512b-7337-4d8e-a8cf-880f4f08fd08"
+    private val videoId = "C3XLe1CCcOq"
+    private val orgCode = "demoveranda"
     lateinit var player: TpStreamPlayer
     private var parameters : TpInitParams? = null
     val TAG = "PlayerActivity"
@@ -27,6 +27,7 @@ class PlayerActivity : AppCompatActivity() {
             val pausedAt = sharedPreference.getLong("pausedAt", 0L)
 
             override fun onInitializationSuccess(player: TpStreamPlayer) {
+                this@PlayerActivity.player = player
                 if (parameters == null){
                     parameters = TpInitParams.Builder()
                         .setVideoId(videoId)
@@ -37,79 +38,18 @@ class PlayerActivity : AppCompatActivity() {
                         .build()
                 }
                 playerFragment.load(parameters!!)
-                this@PlayerActivity.player = player
+                addPlayerListener()
             }
         });
         playerFragment.enableAutoFullScreenOnRotate()
-        playerFragment.playbackStateListener = object : TPPlayerListener {
-            override fun onDeviceInfoChanged(deviceInfo: DeviceInfo) {
-                Log.d(TAG, "onDeviceInfoChanged: ")
-            }
+    }
 
-            override fun onEvents(player: TpStreamPlayer?, events: Player.Events) {
-                Log.d(TAG, "onEvents: ")
-            }
-
-            override fun onIsLoadingChanged(loading: Boolean) {
-                Log.d(TAG, "onIsLoadingChanged: ")
-            }
-
-            override fun onIsPlayingChanged(playing: Boolean) {
-                Log.d(TAG, "onIsPlayingChanged: ")
-            }
-
-            override fun onMetadata(metadata: Metadata) {
-                Log.d(TAG, "onMetadata: ")
-            }
-
-            override fun onPlayWhenReadyChanged(playWhenReady: Boolean, reason: Int) {
-                Log.d(TAG, "onPlayWhenReadyChanged: ")
-            }
-
+    private fun addPlayerListener(){
+        player.setListener( object : TPStreamPlayerListener {
             override fun onPlaybackStateChanged(playbackState: Int) {
-                /*
-                * Playback states can be compared with TpStreamPlayer.PLAYBACK_STATE
-                * */
-                Log.d(TAG, "onPlaybackStateChanged: ")
+                Log.d(TAG, "onPlaybackStateChanged: $playbackState")
             }
-
-            override fun onPlayerError(error: PlaybackException) {
-                Log.d(TAG, "onPlayerError: ")
-            }
-
-            override fun onPlayerErrorChanged(error: PlaybackException?) {
-                Log.d(TAG, "onPlayerErrorChanged: ")
-            }
-
-            override fun onPositionDiscontinuity(
-                oldPosition: Player.PositionInfo,
-                newPosition: Player.PositionInfo,
-                reason: Int
-            ) {
-                Log.d(TAG, "onPositionDiscontinuity: ")
-            }
-
-            override fun onSeekBackIncrementChanged(seekBackIncrementMs: Long) {
-                Log.d(TAG, "onSeekBackIncrementChanged: ")
-            }
-
-            override fun onSeekForwardIncrementChanged(seekForwardIncrementMs: Long) {
-                Log.d(TAG, "onSeekForwardIncrementChanged: ")
-            }
-
-            override fun onTimelineChanged(timeline: Timeline, reason: Int) {
-                Log.d(TAG, "onTimelineChanged: ")
-            }
-
-            override fun onTracksChanged(tracks: Tracks) {
-                Log.d(TAG, "onTracksChanged: ")
-            }
-
-            override fun onVideoSizeChanged(videoSize: VideoSize) {
-                Log.d(TAG, "onVideoSizeChanged: ")
-            }
-
-        }
+        })
     }
 
     override fun onBackPressed() {